+++ src/stores/goal.ts
@@ -1,10 +1,10 @@
 import { defineStore } from 'pinia';
 
-import { cmpGoal } from 'src/lib/goal.ts';
-import { getGoals, Goal } from 'src/lib/api/goal.ts';
+import { cmpGoalByProgress } from 'src/lib/goal.ts';
+import { getGoals, GoalWithAchievement } from 'src/lib/api/goal.ts';
 
 export const useGoalStore = defineStore('goal', {
-  state: () : { goals: Goal[] | null; } => {
+  state: () : { goals: GoalWithAchievement[] | null; } => {
     return { goals: null };
   },
   getters: {
@@ -20,7 +20,7 @@ export const useGoalStore = defineStore('goal', {
     async populate(force = false) {
       if(force || this.goals === null) {
         const goals = await getGoals();
-        this.goals = goals.sort(cmpGoal);
+        this.goals = goals.sort(cmpGoalByProgress);
       }
     }
   },